pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tubert"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "transformers>=5.0.0",
    "numpy>=2.4.4",
    "pyaudio>=0.2.14",
    "vosk>=0.3.44",
    "soundfile>=0.13.1",
    "Pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "pandas>=3.0.2",
    "tqdm>=4.67.1",
    "torchmetrics>=1.8.2",
]

[tool.setuptools.packages.find]
where = ["."]
